Many additions and some bug fixes
- Vcent (\vcenter), Acc (\accent), Over (\overline), and Under
(\underline) atoms have all now been fully implemented. The only atom
type that's left is Rad (\radical) (accents took especially long to make
them look not super ugly when they're on top of characters).
- A lot of new commands, most of them accent macros since that's finally
been implemented. The \skew command is probably the most important since
it's pretty important for accents.
- When you input a string of TeX (`fontTeX.render') and render it, the
parts of the string that aren't enclosed in TeX delimiters ($) will now
be parsed and rendered as HTML. This lets you input a string like "<span
style="color:blue"> $ ... $ </span>" to color the TeX. Any HTML inside
the $ delimiters will not be rendered as HTML (they'll be treated as
literal characters).
- The mathcodes of characters are now editable via the \mathcode
command. This command lets you control spacing around atoms (by changing
their families similar to how \mathord and related commands work) and
allows you to make (sort of) active characters (in plain TeX, the
apostrophe character is the only one of its kind).
- Fractions are rendered at the right height now (at the same height as
what vertical-align: middle uses). Delimiter heights are also calculated
using this new reference for the middle of the line, so their sizes look
more accurate in relation to their contents.
- Fractions wrapped in a group (like "{1\over1}") will now render as an
Inner atom despite being in a group that would normally evaluate to an
Ord atom. This is copied directly from TeX's normal behavior.
- Catcode and mathcode registers (retrieved via \catcode and \mathcode)
now have an upper and lower bound. For example, if you try to set a
character's catcode to anything greater than 15, the assignment will
fail (more specifically, the assignment will just be ignored). The range
for catcodes is [0,15] while mathcodes have a range of [0,8].
- The \obeyspaces command has now been actually implemented. Previously,
it basically did nothing.
- Subscripts and superscripts should now appear all the time.
Previously, if they followed a token like "{" (one that opened a new
group), they would disappear with the curly bracket. Now, they should be
rendered no matter which atom precedes them.
- The apostrophe character no longer evaluates to "^{\prime}" since it
ends up not looking like an apostrophe the way it does in TeX.
- Kerns have been improved so their spacing is more predictable.
Previously, mixing negative and positive kerns would lead to unwanted
line wrapping.
- Fractions with huge denominators don't look weird now.
- Atoms that are rendered in a different font-size than their
neighboring atoms now correctly affect the height and depth of their
parents (for example, when you use \displaystyle in a box meant to be
rendered in scripscript style).
- When a subscript or superscript character is left without a token to
make into a script (like if you ended the TeX with a "_"), the
sub/superscript character will not be marked invalid and be rendered in
red so you know that something is wrong with your syntax.
- \mathchar and \mathchardef work now.
- \accent is now evaluated after commands like \mathord. That lets you
do something like '\accent"B4\mathord{+}' to make an accented Ord atom
with a "+" nucleus instead of having to group the '\mathord{+}' in its
own group.
- Tokens are simplified after being parsed now, which means less
unnecessary creating of elements, which means faster atoms are turned
into HTML faster.
- The \show command now correctly shows active character macros.
- There was a bug when determining the height and depth of Op atoms
(\mathop) with limits. Now, heights and depths should be calculated
correctly.
- There are two ways TeX is recognized in a string. If it starts with
just one $, it should also end in just $, If it starts with two $, the
TeX should also end in two $. Now, if the TeX was started with two $$,
and a partner-less $ is found in the middle of the TeX, it's marked as
invalid instead of just being allowed to exist.
- A problem with parsing mu dimensions has been fixed. Previously, if
you meant to \mkern 1mu, you might end up getting 65536 mu instead of
just the 1.
- Some unnecessary/commented out code has now been removed.