Fixed two bugs and added a minified version
In some browsers, the width of a character wasn't being determined
correctly (fontTeX.fontDimen.widthOf) because the font size was set to
1px. Instead of returning a decimal number that could be applied to em
units, it would return either a 0 or 1 (rounded), which threw off all
the spacing for characters. That's been fixed so that it will be at
least accurate to 0.02.
There was a bug where the autoupdater would get rid of certain
informtion about atoms. Accents and radicals for example would lose
their extra information and just be rendered as a regular Ord atom.
That's been fixed so that the auto updater should result in the same
rendered TeX, just in another font size/family.
A minified version has been added (using the Google's closure compiler).
New \not and \cong macros.